06.04

+++ rentacar/src/main/java/turkcell/rentacar/entities/concretes/Invoice.java
@@ -9,6 +9,7 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
+import javax.persistence.OneToOne;
 import javax.persistence.Table;
 
 import lombok.AllArgsConstructor;
@@ -33,7 +34,7 @@ public class Invoice {
 	private LocalDate createdDate;
 	
 	@Column(name = "rental_days_value")
-	private double rentalDaysValue;		
+	private long rentalDaysValue;		
 	
 	@Column(name = "rental_date")
 	private LocalDate rentalDate;
